Revert "perf: Changed every variable to short and bytes (as they cannot exceed them), changed string to char[]"
This reverts commit d0e4216b17bfb7dc332b46e027f851dda4556f88.

+++ HumanCalculator/Score.cs
@@ -16,7 +16,7 @@ namespace HumanCalculator
         {
             this.Time = new Time();
         }
-        public void AddTime(byte seconds)
+        public void AddTime(double seconds)
         {
             this.Time.Times[currentLoop]=seconds;
             currentLoop++;
